Added period information to timetable
This will be necessary for the SQLite data to cope with doubles.

Additionally, the file test/db/lessons_and_notes.db is included
in this commit. No test cases use it yet; it was created automatically
upon instantiating the test Database object. It will surely play a
role soon though.